Lectures, featured lectures and file thumbnails
Task model has been extended with is_lecture, lecture_folder and
lecture_video_url to support lectures. All lectures are actually an
extension of files.

List of featured lectures is shown in homepage. Lecture is featured if
it is contained in the folder with ID equal to
settings.FEATURED_LECTURES_FOLDER_ID.